+++ minijinja/src/compiler/parser.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeSet;
 use std::fmt;
 
 use crate::compiler::ast::{self, Spanned};
-use crate::compiler::lexer::tokenize;
+use crate::compiler::lexer::{tokenize, SyntaxConfig};
 use crate::compiler::tokens::{Span, Token};
 use crate::error::{Error, ErrorKind};
 use crate::value::Value;
@@ -95,8 +95,8 @@ struct TokenStream<'a> {
 
 impl<'a> TokenStream<'a> {
     /// Tokenize a template
-    pub fn new(source: &'a str, in_expr: bool) -> TokenStream<'a> {
-        let mut iter = Box::new(tokenize(source, in_expr)) as Box<dyn Iterator<Item = _>>;
+    pub fn new(source: &'a str, in_expr: bool, syntax: SyntaxConfig) -> TokenStream<'a> {
+        let mut iter = Box::new(tokenize(source, in_expr, syntax)) as Box<dyn Iterator<Item = _>>;
         let current = iter.next();
         TokenStream {
             iter,
@@ -216,9 +216,9 @@ macro_rules! with_recursion_guard {
 }
 
 impl<'a> Parser<'a> {
-    pub fn new(source: &'a str, in_expr: bool) -> Parser<'a> {
+    pub fn new(source: &'a str, in_expr: bool, syntax: SyntaxConfig) -> Parser<'a> {
         Parser {
-            stream: TokenStream::new(source, in_expr),
+            stream: TokenStream::new(source, in_expr, syntax),
             in_macro: false,
             blocks: BTreeSet::new(),
             depth: 0,
@@ -1091,7 +1091,17 @@ impl<'a> Parser<'a> {
 }
 
 /// Parses a template
+#[cfg(feature = "unstable_machinery")]
 pub fn parse<'source>(source: &'source str, filename: &str) -> Result<ast::Stmt<'source>, Error> {
+    parse_with_syntax(source, filename, Default::default())
+}
+
+/// Parses a template with a specific syntax
+pub fn parse_with_syntax<'source>(
+    source: &'source str,
+    filename: &str,
+    syntax: SyntaxConfig,
+) -> Result<ast::Stmt<'source>, Error> {
     // we want to chop off a single newline at the end.  This means that a template
     // by default does not end in a newline which is a useful property to allow
     // inline templates to work.  If someone wants a trailing newline the expectation
@@ -1104,7 +1114,7 @@ pub fn parse<'source>(source: &'source str, filename: &str) -> Result<ast::Stmt<
         source = &source[..source.len() - 1];
     }
 
-    let mut parser = Parser::new(source, false);
+    let mut parser = Parser::new(source, false, syntax);
     parser.parse().map_err(|mut err| {
         if err.line().is_none() {
             err.set_filename_and_span(filename, parser.stream.last_span())
@@ -1114,8 +1124,8 @@ pub fn parse<'source>(source: &'source str, filename: &str) -> Result<ast::Stmt<
 }
 
 /// Parses an expression
-pub fn parse_expr(source: &str) -> Result<ast::Expr<'_>, Error> {
-    let mut parser = Parser::new(source, true);
+pub fn parse_expr(source: &str, syntax: SyntaxConfig) -> Result<ast::Expr<'_>, Error> {
+    let mut parser = Parser::new(source, true, syntax);
     parser
         .parse_expr()
         .and_then(|result| {