+++ minijinja/src/environment.rs
@@ -6,6 +6,7 @@ use std::sync::Arc;
 use serde::Serialize;
 
 use crate::compiler::codegen::CodeGenerator;
+use crate::compiler::lexer::SyntaxConfig;
 use crate::compiler::parser::parse_expr;
 use crate::error::{attach_basic_debug_info, Error, ErrorKind};
 use crate::expression::Expression;
@@ -20,7 +21,7 @@ type TemplateMap<'source> = BTreeMap<&'source str, Arc<CompiledTemplate<'source>
 
 #[derive(Clone)]
 enum Source<'source> {
-    Borrowed(TemplateMap<'source>),
+    Borrowed(TemplateMap<'source>, SyntaxConfig),
     #[cfg(feature = "source")]
     Owned(crate::source::Source),
 }
@@ -28,7 +29,7 @@ enum Source<'source> {
 impl<'source> fmt::Debug for Source<'source> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Borrowed(tmpls) => fmt::Debug::fmt(&BTreeMapKeysDebug(tmpls), f),
+            Self::Borrowed(tmpls, _) => fmt::Debug::fmt(&BTreeMapKeysDebug(tmpls), f),
             #[cfg(feature = "source")]
             Self::Owned(arg0) => fmt::Debug::fmt(arg0, f),
         }
@@ -101,7 +102,7 @@ impl<'source> Environment<'source> {
     /// [`empty`](Environment::empty) method.
     pub fn new() -> Environment<'source> {
         Environment {
-            templates: Source::Borrowed(Default::default()),
+            templates: Source::Borrowed(Default::default(), Default::default()),
             filters: defaults::get_builtin_filters(),
             tests: defaults::get_builtin_tests(),
             globals: defaults::get_globals(),
@@ -121,7 +122,7 @@ impl<'source> Environment<'source> {
     /// logic for auto escaping configured.
     pub fn empty() -> Environment<'source> {
         Environment {
-            templates: Source::Borrowed(Default::default()),
+            templates: Source::Borrowed(Default::default(), Default::default()),
             filters: Default::default(),
             tests: Default::default(),
             globals: Default::default(),
@@ -150,8 +151,12 @@ impl<'source> Environment<'source> {
     )]
     pub fn add_template(&mut self, name: &'source str, source: &'source str) -> Result<(), Error> {
         match self.templates {
-            Source::Borrowed(ref mut map) => {
-                let compiled_template = ok!(CompiledTemplate::from_name_and_source(name, source));
+            Source::Borrowed(ref mut map, ref syntax) => {
+                let compiled_template = ok!(CompiledTemplate::from_name_and_source_with_syntax(
+                    name,
+                    source,
+                    syntax.clone()
+                ));
                 map.insert(name, Arc::new(compiled_template));
                 Ok(())
             }
@@ -163,7 +168,7 @@ impl<'source> Environment<'source> {
     /// Removes a template by name.
     pub fn remove_template(&mut self, name: &str) {
         match self.templates {
-            Source::Borrowed(ref mut map) => {
+            Source::Borrowed(ref mut map, _) => {
                 map.remove(name);
             }
             #[cfg(feature = "source")]
@@ -188,7 +193,7 @@ impl<'source> Environment<'source> {
     /// ```
     pub fn get_template(&self, name: &str) -> Result<Template<'_>, Error> {
         let compiled = match &self.templates {
-            Source::Borrowed(ref map) => {
+            Source::Borrowed(ref map, _) => {
                 ok!(map.get(name).ok_or_else(|| Error::new_not_found(name)))
             }
             #[cfg(feature = "source")]
@@ -251,7 +256,11 @@ impl<'source> Environment<'source> {
     }
 
     fn _render_str(&self, name: &str, source: &str, root: Value) -> Result<String, Error> {
-        let compiled = ok!(CompiledTemplate::from_name_and_source(name, source));
+        let compiled = ok!(CompiledTemplate::from_name_and_source_with_syntax(
+            name,
+            source,
+            self._syntax_config().clone()
+        ));
         let mut rv = String::with_capacity(compiled.buffer_size_hint);
         Vm::new(self)
             .eval(
@@ -396,6 +405,38 @@ impl<'source> Environment<'source> {
         self.fuel
     }
 
+    /// Sets the syntax for the environment.
+    ///
+    /// Note that when `source` is used, the syntax is held on the underlying source
+    /// which means that the actual source needs to have it's syntax changed.
+    ///
+    /// See [`Syntax`](crate::Syntax) for more information.
+    #[cfg(feature = "custom_syntax")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "custom_syntax")))]
+    pub fn set_syntax(&mut self, syntax: crate::custom_syntax::Syntax) -> Result<(), Error> {
+        match self.templates {
+            Source::Borrowed(_, ref mut syn) => *syn = ok!(syntax.compile()),
+            #[cfg(feature = "source")]
+            Source::Owned(ref mut source) => ok!(source.set_syntax(syntax)),
+        };
+        Ok(())
+    }
+
+    /// Returns the current syntax.
+    #[cfg(feature = "custom_syntax")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "custom_syntax")))]
+    pub fn syntax(&self) -> &crate::custom_syntax::Syntax {
+        &self._syntax_config().syntax
+    }
+
+    fn _syntax_config(&self) -> &SyntaxConfig {
+        match self.templates {
+            Source::Borrowed(_, ref syn) => syn,
+            #[cfg(feature = "source")]
+            Source::Owned(ref source) => source._syntax_config(),
+        }
+    }
+
     /// Sets the template source for the environment.
     ///
     /// This helps when working with dynamically loaded templates.  The
@@ -419,7 +460,7 @@ impl<'source> Environment<'source> {
     #[cfg_attr(docsrs, doc(cfg(feature = "source")))]
     pub fn source(&self) -> Option<&crate::source::Source> {
         match self.templates {
-            Source::Borrowed(_) => None,
+            Source::Borrowed(..) => None,
             Source::Owned(ref source) => Some(source),
         }
     }
@@ -429,7 +470,7 @@ impl<'source> Environment<'source> {
     #[cfg_attr(docsrs, doc(cfg(feature = "source")))]
     pub fn source_mut(&mut self) -> Option<&mut crate::source::Source> {
         match self.templates {
-            Source::Borrowed(_) => None,
+            Source::Borrowed(..) => None,
             Source::Owned(ref mut source) => Some(source),
         }
     }
@@ -445,7 +486,7 @@ impl<'source> Environment<'source> {
     }
 
     fn _compile_expression(&self, expr: &'source str) -> Result<Expression<'_, 'source>, Error> {
-        let ast = ok!(parse_expr(expr));
+        let ast = ok!(parse_expr(expr, self._syntax_config().clone()));
         let mut gen = CodeGenerator::new("<expression>", expr);
         gen.compile_expr(&ast);
         let (instructions, _) = gen.finish();