+++ minijinja/src/template.rs
@@ -5,7 +5,8 @@ use serde::Serialize;
 
 use crate::compiler::codegen::CodeGenerator;
 use crate::compiler::instructions::Instructions;
-use crate::compiler::parser::parse;
+use crate::compiler::lexer::SyntaxConfig;
+use crate::compiler::parser::parse_with_syntax;
 use crate::environment::Environment;
 use crate::error::{attach_basic_debug_info, Error, ErrorKind};
 use crate::output::{Output, WriteWrapper};
@@ -182,21 +183,35 @@ impl<'env> fmt::Debug for CompiledTemplate<'env> {
 
 impl<'source> CompiledTemplate<'source> {
     /// Creates a compiled template from name and source.
+    #[cfg(feature = "unstable_machinery")]
     pub fn from_name_and_source(
         name: &'source str,
         source: &'source str,
     ) -> Result<CompiledTemplate<'source>, Error> {
-        attach_basic_debug_info(Self::_from_name_and_source_impl(name, source), source)
+        Self::from_name_and_source_with_syntax(name, source, Default::default())
     }
 
-    fn _from_name_and_source_impl(
+    /// Creates a compiled template from name and source using the given settings.
+    pub fn from_name_and_source_with_syntax(
         name: &'source str,
         source: &'source str,
+        syntax: SyntaxConfig,
+    ) -> Result<CompiledTemplate<'source>, Error> {
+        attach_basic_debug_info(
+            Self::_from_name_settings_and_source_with_syntax_impl(name, source, syntax),
+            source,
+        )
+    }
+
+    fn _from_name_settings_and_source_with_syntax_impl(
+        name: &'source str,
+        source: &'source str,
+        syntax: SyntaxConfig,
     ) -> Result<CompiledTemplate<'source>, Error> {
         // the parser/compiler combination can create constants in which case
         // we can probably benefit from the value optimization a bit.
         let _guard = value::value_optimization();
-        let ast = ok!(parse(source, name));
+        let ast = ok!(parse_with_syntax(source, name, syntax));
         let mut gen = CodeGenerator::new(name, source);
         gen.compile_stmt(&ast);
         let buffer_size_hint = gen.buffer_size_hint();