+++ minijinja/tests/test_templates.rs
@@ -424,3 +424,24 @@ fn test_flattening_sub_item_shielded_print() {
         r#"{"bad": <invalid value: can only flatten structs and maps (got an enum)>}"#
     );
 }
+
+#[test]
+#[cfg(feature = "custom_syntax")]
+fn test_custom_syntax() {
+    let mut env = Environment::new();
+    env.set_syntax(minijinja::Syntax {
+        block_start: "{".into(),
+        block_end: "}".into(),
+        variable_start: "${".into(),
+        variable_end: "}".into(),
+        comment_start: "{*".into(),
+        comment_end: "*}".into(),
+    })
+    .unwrap();
+
+    // this on the other hand is okay
+    let value = env
+        .render_str("{for x in range(3)}${x}{endfor}{* nothing *}", ())
+        .unwrap();
+    assert_eq!(value, r"012");
+}